Monte Carlo Simulation to investigate the optimal number of points to use in the moment conditions.

Runs Monte Carlo Simulation for different values of \(\alpha\) and \(\beta\) and computes a specified number of t-points that minimises the determinant of the asymptotic covariance matrix.

ComputeBest_t(AlphaBetaMatrix = abMat, nb_ts = seq(10, 100, 10),
              alphaReg = 0.001,FastOptim=TRUE,...)

Arguments

AlphaBetaMatrix
Values of the parameter \(\alpha\) and \(\beta\) from which we simulate the data. By default, The values of \(\gamma\) and \(\delta\) are set to 1 and 0 respectively; matrix \(2 \times n\).

nb_ts
Vector of number of t-points to use for the minimisation; default seq(10,100,10).

alphaReg
Value of the regularisation parameter; numeric, default is 0.001.

FastOptim
Logical flag; if set to TRUE, optim with "Nelder-Mead" method is used (fast but not accurate). Otherwise, nlminb is used (more accurate but slower).

…
Other arguments to pass to the optimisation function.

Value

a list containing slots from class Best_t-class corresponding to one value of the parameters \(\alpha\) and \(\beta\).

See also